+++ doc/changes.md
@@ -32,6 +32,10 @@ and terminal emulations that *do not* use the new 256-color mode,
    - RGB values for colors 16 and up can be customised with the new 
      `SET GUI RGBCOLOR INDEX` command 
    - RGB values for colors 16 and up can now be shown with `SHOW GUI /PALETTE`
+ - Support for direct 24-bit color via SGR-38/SGR-48 in K95G on newer versions 
+   of Windows. On older windows releases, the nearest color in the current color
+   palette is used when a 24-bit color is set. In the OS/2 and Windows console
+   versions, the nearest color in the 16-color palette is always used.
  - Support xterm OSC-4 operating system command for changing the color palette
  - Support xterm OSC-104 operating system command for resetting the color palette
  - Support xterm OSC-2 and OSC-l operating system commands for setting the